+++ ionic2/app/pages/pokedetail/pokedetail.page.html
