+++ ionic2/app/theme/app.core.scss
@@ -9,7 +9,7 @@
 
 @import "../pages/map/map.page";
 @import "../pages/about/about.page";
-@import "../pages/wiki/wiki.page";
+@import "../pages/pokedetail/pokedetail.page";
 @import "../pages/search/search.page";
 
 @import "../components/map/map.component";
